move mongodb storage config to util
